Some abs back to Fabs... abs in standalone mode rounds everything to integers =( we sometimes need doubles

+++ src/mqmprob.cpp
@@ -177,7 +177,7 @@ double left_prob(const double r,const MQMMarker markerL,const MQMMarker markerR,
   const double rr = 1.0-r;      // No recombination
   const double rr2 = rr*rr;     //Double Norecombination
  
-  const int recombinations = (int)abs((double)markerL-(double)markerR);    //The number of recombinations
+  const int recombinations = abs(markerL-markerR);    //The number of recombinations
 
   switch (crosstype) {
     case CF2: //F2 cross
@@ -317,7 +317,7 @@ double right_prob_F2(const char markerL, const int j, const MQMMarkerVector imar
   const double r2 = r*r;      //Breeding Logic (see prob_new)
   const double rr = 1.0-r;
   const double rr2 = rr*rr;
-  const int recombinations = (int)abs(markerL-markerR);   //Number of recombinations between markerL and markerR
+  const int recombinations = abs(markerL-markerR);   //Number of recombinations between markerL and markerR
 
   if (is_knownMarker(markerR, CF2)) {   //If we know the next marker we have an answer
     if ((markerL==MH)&&(markerR==MH)) {
@@ -391,7 +391,7 @@ double right_prob_BC(const char markerL, const int j, const MQMMarkerVector imar
   double prob0 = 0.0;                                  		//Internal variable holding the probability AA if the next rightmarker is (Semi) Unknown
   double prob1 = 0.0;                                  		//Internal variable holding the probability H if the next rightmarker is (Semi) Unknown
   const double rr = 1.0-r;                             		//Breeding Logic (see prob_new)
-  const int recombinations = (int)abs(markerL-markerR);    //Number of recombinations between markerL and markerR
+  const int recombinations = abs(markerL-markerR);    //Number of recombinations between markerL and markerR
   if (is_knownMarker(markerR, CBC)) {
     return ((recombinations==0)? rr : r );
   } else {
